+++ src/types/addItemPreviewTypes.ts
@@ -1,0 +1,4 @@
+export interface IPreview {
+  id: string;
+  url: string;
+}
