+++ src/types/addItemTagTypes.ts
@@ -1,0 +1,4 @@
+export interface ITag {
+  id: string;
+  content: string;
+}
